Adding confidentializable support to lkvm
We want to allow the guest to fully boot before transitioning into
a confidential VM.
I introduced a new type of memory segment called "Confidentializable".
The linux tyche driver will cut a segment, duplicate it, and send the
original cut to the domain along the revocation handle for the
duplicated part (that it keeps).
This allows the child to revoke the parent access whenever it wants.